MA-22696 Remove build warnings of NXPCameraHAL and ExtCameraHAL
Most warnings are caused by inaccurate data format matching, which may
result in loss of precision. When using, strictly follow the data type
to fix these warnings.

Reviewed-on: http://androidsource.nxp.com/project/25471
(cherry picked from commit a8932cb642e2070d14bafed25553240e15bc241d)
Reviewed-on: http://androidsource.nxp.com/project/25478

+++ camera/camera-hal/CameraDeviceHWLImpl.cpp
@@ -362,7 +362,7 @@ bool CameraDeviceHwlImpl::StreamCombJudge(const StreamConfiguration &stream_conf
                                           int *pPictureResolutions, int nPictureResolutionCount) {
     for (const auto &stream : stream_config.streams) {
         if (stream.stream_type != google_camera_hal::StreamType::kOutput) {
-            ALOGE("%s: only support stream type output, but it's %d", __func__, stream.stream_type);
+            ALOGE("%s: only support stream type output, but it's %d", __func__, (int)stream.stream_type);
             return false;
         }
